emacs: navigation: abort if mouse leaves or kbd macro ends
Still need to make it for when the user leaves the minibuffer.

One possible solution would be to figure out which buffers have the
goto-line-relative active and store the various variables required in
lists in the order of the recursive edit. This would the user to use
`enable-recursive-minibuffers` but that's a good thing right??