vnc: switch to x11vnc instead of tigervnc due to x11 input problems
tigervnc is not able to translate the keycodes properly thus leading
to the error "x0vncserver - Could not map key event to X11 key
code". This affected almost all keys (only small case letters
worked). No proper fix exists for it, adding -RawKeyboard as a
parameter is proposed as a way of fixing it but this didn't help.