drivers: huion: wacom_settings: map the stylus based on x11 geometry
This ensures that as long as the monitors are in the same position, it
will always work.

Specifying using `HEAD-2` or similar can lead selecting the wrong
monitor since this can vary based on which monitor was connected
first. Specifying using the x11 geometry, it is guaranteed to be the
correct monitor as long as the physical layout of the monitors does
not change.

The best way to handle this would be to create a GUI application that
will generate these commands on the fly. Will be an interesting
challenge.