emacs: lsp-rust-analyzer-cargo-watch-command-toggle & default check
Switch the default cargo watch command to check to make things faster,
clippy is too slow :( even though it is probably the better thing to
have. But to be able to at least have access to clippy in lsp,
function (with binding "C-c l c") to toggle between clippy and check.